pyproject.toml
[tool.poetry]
name = "python_anywhere"
version = "0.1.0"
description = ""
authors = ["klu0926 <lukuoyu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.1.1"
pillow = "^11.2.1"
flask-cors = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

